prepping for prod release- adding dbo

+++ Source/Rio.Database/Scripts/clean up test accounts.sql
@@ -1,3 +1,3 @@
-delete from accountuser where accountid in (select accountid from account where updatedate is null)
+delete from dbo.accountuser where accountid in (select accountid from dbo.account where updatedate is null)
 
-delete from account where updatedate is null
+delete from dbo.account where updatedate is null
